Fixing x64 compilation, silence some dgbprint and other improvements

+++ wrappers/base/ntext_wrapper/rtl/heap.c
@@ -41,7 +41,7 @@ RtlpFreeMemory(
 PVOID 
 NTAPI 
 RtlpAllocateMemory( 	
-	ULONG  	Bytes,
+	SIZE_T  	Bytes,
 	ULONG  	Tag 
 ) 		
 {
